requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
azure-communication-email==1.0.0
